requirements.txt
agno
yfinance
duckduckgo-search
groq
python-dotenv